Updated simple_interest.py with local changes

+++ python_introduction/simple_interest.py
@@ -4,7 +4,7 @@ rate = 0.05
 time =3
 
 #Calculate simple_interest
-simple_interest = float(principal*rate*time)
+simple_interest = principal*rate*time
 
 #Printing results
 print('The simple interest is:', simple_interest)
